BUGFIX: Bloqueo de eliminacion si una habitacion tiene camillas asociadas a ella

+++ src/Controller/HabitacionController.php
@@ -458,27 +458,45 @@ class HabitacionController extends AbstractController
      */
     public function delete(Request $request, Habitacion $habitacion, Clinica $clinica, Security $AuthUser): Response
     {
-        //VALIDACION DE REGISTROS UNICAMENTE DE MI CLINICA SI NO SOY ROLE_SA
-        if($AuthUser->getUser()->getRol()->getNombreRol() != 'ROLE_SA'){
-            if($AuthUser->getUser()->getClinica()->getId() == $clinica->getId() && $AuthUser->getUser()->getClinica()->getId() == $habitacion->getSala()->getClinica()->getId() ){
-                if ($this->isCsrfTokenValid('delete'.$habitacion->getId(), $request->request->get('_token'))) {
-                    $entityManager = $this->getDoctrine()->getManager();
-                    $entityManager->remove($habitacion);
-                    $entityManager->flush();
-                }
-                $this->addFlash('success','Habitación eliminada con éxito');
-                return $this->redirectToRoute('habitacion_index',[
-                    'clinica' => $clinica->getId(),
-                ]);
+
+        $em = $this->getDoctrine()->getManager();
+        $RAW_QUERY = "SELECT COUNT(*) as cuenta FROM habitacion as h, camilla as c WHERE 
+        c.habitacion_id = h.id AND
+        h.id = ".$habitacion->getId().";";
+        $statement = $em->getConnection()->prepare($RAW_QUERY);
+        $statement->execute();
+        $validacionBloqueoEliminar = $statement->fetchAll();
+        if((int) $validacionBloqueoEliminar[0]["cuenta"] >= 1){
+            $this->addFlash('notice','Para borrar la habitacion verifique que este no tenga camillas asociadas a el');
+            if($AuthUser->getUser()->getRol()->getNombreRol() == 'ROLE_SA'){
+                return $this->redirectToRoute('habitacion_index',array('clinica'=>$clinica->getId()));
             }else{
-                $this->addFlash('fail','Error, este registro puede que no exista o no le pertenece');
-                return $this->redirectToRoute('habitacion_index',array('clinica'=>$AuthUser->getUser()->getClinica()->getId()));
+                return $this->redirectToRoute('habitacion_index',array('clinica'=>$AuthUser->getUser()->getClinica()->getId()));  
             }
-        }
-        if ($this->isCsrfTokenValid('delete'.$habitacion->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($habitacion);
-            $entityManager->flush();
+            
+        }else{
+            //VALIDACION DE REGISTROS UNICAMENTE DE MI CLINICA SI NO SOY ROLE_SA
+            if($AuthUser->getUser()->getRol()->getNombreRol() != 'ROLE_SA'){
+                if($AuthUser->getUser()->getClinica()->getId() == $clinica->getId() && $AuthUser->getUser()->getClinica()->getId() == $habitacion->getSala()->getClinica()->getId() ){
+                    if ($this->isCsrfTokenValid('delete'.$habitacion->getId(), $request->request->get('_token'))) {
+                        $entityManager = $this->getDoctrine()->getManager();
+                        $entityManager->remove($habitacion);
+                        $entityManager->flush();
+                    }
+                    $this->addFlash('success','Habitación eliminada con éxito');
+                    return $this->redirectToRoute('habitacion_index',[
+                        'clinica' => $clinica->getId(),
+                    ]);
+                }else{
+                    $this->addFlash('fail','Error, este registro puede que no exista o no le pertenece');
+                    return $this->redirectToRoute('habitacion_index',array('clinica'=>$AuthUser->getUser()->getClinica()->getId()));
+                }
+            }
+            if ($this->isCsrfTokenValid('delete'.$habitacion->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($habitacion);
+                $entityManager->flush();
+            }
         }
         $this->addFlash('success','Habitación eliminada con éxito');
         return $this->redirectToRoute('habitacion_index',[
